+++ app/Models/Account/Account.php
@@ -22,6 +22,7 @@ use App\Models\Contact\Message;
 use App\Models\Contact\Document;
 use App\Models\Contact\Reminder;
 use App\Models\Contact\LifeEvent;
+use App\Models\Instance\AuditLog;
 use App\Services\User\CreateUser;
 use App\Models\Contact\Occupation;
 use Illuminate\Support\Facades\DB;
@@ -477,6 +478,16 @@ class Account extends Model
     public function occupations()
     {
         return $this->hasMany(Occupation::class);
+    }
+
+    /**
+     * Get the Audit log records associated with the account.
+     *
+     * @return HasMany
+     */
+    public function auditLogs()
+    {
+        return $this->hasMany(AuditLog::class);
     }
 
     /**