+++ tests/Unit/Models/AccountTest.php
@@ -20,6 +20,7 @@ use App\Models\Account\Activity;
 use App\Models\Contact\Document;
 use App\Models\Contact\Reminder;
 use App\Models\Contact\LifeEvent;
+use App\Models\Instance\AuditLog;
 use App\Models\Account\Invitation;
 use App\Models\Contact\Occupation;
 use Illuminate\Support\Facades\DB;
@@ -251,6 +252,16 @@ class AccountTest extends FeatureTestCase
             'account_id' => $account->id,
         ]);
         $this->assertTrue($account->occupations()->exists());
+    }
+
+    /** @test */
+    public function it_has_many_logs()
+    {
+        $account = factory(Account::class)->create([]);
+        factory(AuditLog::class)->create([
+            'account_id' => $account->id,
+        ]);
+        $this->assertTrue($account->auditLogs()->exists());
     }
 
     /** @test */