+++ tests/Unit/Models/ContactTest.php
@@ -16,6 +16,7 @@ use App\Models\Account\Activity;
 use App\Models\Contact\Document;
 use App\Models\Contact\Reminder;
 use App\Models\Contact\LifeEvent;
+use App\Models\Instance\AuditLog;
 use App\Models\Contact\Occupation;
 use App\Models\Contact\Conversation;
 use App\Models\Instance\SpecialDate;
@@ -132,6 +133,16 @@ class ContactTest extends FeatureTestCase
             'contact_id' => $contact->id,
         ]);
         $this->assertTrue($contact->occupations()->exists());
+    }
+
+    /** @test */
+    public function it_has_many_logs()
+    {
+        $contact = factory(Contact::class)->create();
+        factory(AuditLog::class, 2)->create([
+            'about_contact_id' => $contact->id,
+        ]);
+        $this->assertTrue($contact->logs()->exists());
     }
 
     /** @test */