+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### New features:
 
+* Add support for audit logs on a contact page
+* Add support for audit logs in the Settings page
 * Add vue data validations
 * Add ability to edit activities
 * Associate a photo to a gift